refactor(telegram): move keyboard builders to keyboards.rs, verbatim (#471 phase 1)
Fifth mechanical cut: the tool-approval callback (Yes/Always/No
keyboard prompt), the /cd directory browser keyboard and the profile
picker keyboard move to keyboards.rs. Pure move, glob re-export keeps
all paths stable.
